+++ Changelog.md
@@ -1,5 +1,8 @@
 ## Changelog
 
+### Next
+- A new completion callback can be set on `Mock` to use for expectation fulfilling once a `Mock` is used.
+
 ### 1.3.0
 - Updated to Swift 4.2