+++ Changelog.md
@@ -3,7 +3,9 @@
 ### Next
 - A new completion callback can be set on `Mock` to use for expectation fulfilling once a `Mock` is used.
 - Updated to Swift 5.0
+- Only dispatch to the background queue if needed
 - Correctly handle cancellation of delayed responses
+- Adding and reading mocks is now thread safe by using a Dispatch Semaphore
 
 ### 1.3.0
 - Updated to Swift 4.2