+++ Changelog.md
@@ -1,6 +1,8 @@
 ## Changelog
 
 ### Next
+
+### 2.0.0
 - A new completion callback can be set on `Mock` to use for expectation fulfilling once a `Mock` is completed.
 - A new onRequest callback can be set on `Mock` to use for expectation fulfilling once a `Mock` is requested.
 - Updated to Swift 5.0